refactor: improve task execution time tracking accuracy
- Change execution_time_ms from int to float for millisecond precision
- Store actual execution time reported by instance separately
- Use instance-reported time over timestamp-based calculation
- Add set_execution_time() method for explicit time setting

This allows for more accurate execution time tracking when instances
report precise timing measurements, while maintaining timestamp-based
fallback calculation for compatibility.

+++ scheduler/src/model.py
@@ -189,7 +189,7 @@ class TaskDetailInfo(BaseModel):
     result: Optional[Dict[str, Any]] = None
     error: Optional[str] = None
     timestamps: TaskTimestamps
-    execution_time_ms: Optional[int] = None
+    execution_time_ms: Optional[float] = None
 
 
 class TaskDetailResponse(BaseModel):
@@ -289,7 +289,7 @@ class WSTaskResultMessage(BaseModel):
     result: Optional[Dict[str, Any]] = None
     error: Optional[str] = None
     timestamps: TaskTimestamps
-    execution_time_ms: Optional[int] = None
+    execution_time_ms: Optional[float] = None
 
 
 class WSErrorMessage(BaseModel):
